florida_church_agent/requirements.txt
beautifulsoup4>=4.12.3
httpx>=0.27.0
lxml>=5.2.2
openpyxl>=3.1.5
pandas>=2.2.2
playwright>=1.52.0
pydantic>=2.7.1
python-dotenv>=1.0.1
rapidfuzz>=3.9.3
requests>=2.32.3
tenacity>=8.4.2
tqdm>=4.66.4
typer>=0.12.3
pydantic-settings>=2.3.1
